+++ lib/extensions/formattable/coded_value.rb
@@ -1,0 +1,17 @@
+# frozen_string_literal: true
+
+module ConnectProto
+  module Extensions
+    module Formattable
+      module CodedValue
+        def to_hl7
+          [
+            value.presence,
+            description.presence,
+            code_set.presence
+          ].compact.join('^').squeeze(' ').strip
+        end
+      end
+    end
+  end
+end