+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Added `MedicationAdministration`
 - Added `repeated MedicationAdministration` to `Order.Order`
+- Added `Formattable::CodedValue` extension
 
 ## 0.10.1